+++ test/benchmark/runner.js
@@ -30,6 +30,11 @@ const filenamesToOpen = [];
  * @returns Promise<any>
  */
 async function run(filename, options) {
+  const fhirpathOptions = {
+    ...(options.mathMode ? {preciseMath : options.mathMode === 'precise'} : {} )
+  };
+
+
   const suites = [].concat(...require('./'+ filename)({
     benny,
     open,
@@ -45,7 +50,8 @@ async function run(filename, options) {
     bigItemsCopy,
     smallItems,
     smallItemsCopy,
-    options
+    options,
+    fhirpathOptions
   }));
   for (const suite of suites) {
     {
@@ -57,7 +63,7 @@ async function run(filename, options) {
               this,
               previous_fhirpath,
               previous_r4_model,
-              previous_fhirpath.compile(suite.expression, previous_r4_model)
+              previous_fhirpath.compile(suite.expression, previous_r4_model, fhirpathOptions)
             )
           )
         );
@@ -68,7 +74,7 @@ async function run(filename, options) {
               this,
               current_fhirpath,
               current_r4_model,
-              current_fhirpath.compile(suite.expression, current_r4_model)
+              current_fhirpath.compile(suite.expression, current_r4_model, fhirpathOptions)
             )
           )
         );
@@ -98,7 +104,7 @@ async function run(filename, options) {
 
 // Retrieves options from the main process and runs benchmarks
 process.on('message', async (options) => {
-  const filenames = options.tests.split(',');
+  const filenames = options.tests;
   for (const filename of filenames) {
     await run(filename, options);
   }